Add pyQT - option

+++ install.sh
@@ -1,8 +1,44 @@
 #!/usr/bin/env bash
 
+app_path_name="/opt/Vpn_Monitor"
 
-# check if packages exists
-cat > check_list.py <<EOL
+check_install () {
+
+    read -p "Proceed ? (Y/n) " response;
+
+    if [[ "${response,}" == y* || -z $response ]];then
+        sudo mkdir $app_path_name
+    else
+        echo "install.sh exited.";
+        exit 0;
+    fi
+
+}
+
+# getting the option of app version.
+echo "write from [gi / qt] the version of app you want to install";
+echo "Note: [gi] is the default choice ..".
+read -p "[gi / qt] ? " install_option;
+
+
+if [[ $install_option == "qt" ]]; then
+    echo "installing QT version.."
+    check_install
+    running_script="ExpressVpn_pyqt.py"
+
+    pip3 install -r requirements.txt
+
+    cp vpn_working.png icon.ico
+    sudo cp ExpressVpn_pyqt.py icon.ico $app_path_name
+
+# case installing gi version
+else
+    echo "installing gi version"
+    check_install
+    running_script="ExpressVpn_Monitor.py"
+
+    # check if packages exists
+    cat > check_list.py <<EOL
 
 import os
 got = os.popen('apt list --installed').read()
@@ -13,23 +49,24 @@ for i in ["python3-gi", "gir1.2-appindicator3-0.1"]:
 
 EOL
 
-pak=`python3 check_list.py`
-rm check_list.py
+    pak=`python3 check_list.py`
+    rm check_list.py
 
-if [[ $pak == "install" ]]; then
-    sudo apt update && sudo apt install gir1.2-appindicator3-0.1 python3-gi -y
+    if [[ $pak == "install" ]]; then
+        sudo apt update && sudo apt install gir1.2-appindicator3-0.1 python3-gi -y
+    fi
+
+# Ending (choosing option)
 fi
+
 
 # setting paths
 user_name=`whoami`
 service_file_name="startExpressVpn_Monitor.service"
-app_path_name="/opt/Vpn_Monitor/"
 
-# create the app folder
-sudo mkdir /opt/Vpn_Monitor
 
 # copying the needed files into app folder
-sudo cp ExpressVpn_Monitor.py vpn_working.png wrong.png $app_path_name
+sudo cp ExpressVpn_Monitor.py common_functions.py vpn_working.png wrong.png $app_path_name
 
 # checking if the needed systemd dir exists
 if [ ! -d "/home/$user_name/.config/systemd/user" ];then
@@ -49,14 +86,14 @@ Description=ExpressVpn Monitoring
 
 [Service]
 Type=simple
-ExecStart=/opt/Vpn_Monitor/ExpressVpn_Monitor.py 
+ExecStart=$app_path_name/$running_script
 
 [Install]
 WantedBy=default.target
 EOL
 
-# binding the service to gnome-session
-if [[ `systemctl --user is-active gnome-session-manager@ubuntu.service ` == "active" ]]; then 
+# binding the service to gnome-session if available
+if [[ `systemctl --user is-active gnome-session-manager@ubuntu.service ` == "active" ]]; then
 	sed -i "3 a ; NOTE : THIS LINE CAN BE IGNORED .. it's only usage to make sure that the service closes BEFORE being forced to fail when the pc is rebooting or shutting down and losing it's gui resources." $service_file_name
 	sed -i "4 a BindsTo=gnome-session-manager@ubuntu.service\n " $service_file_name
 fi
@@ -64,15 +101,8 @@ fi
 # copy service file
 cp $service_file_name /home/$user_name/.config/systemd/user/
 
-# make the (app & service) files executable
-sudo chmod +x $app_path_name/ExpressVpn_Monitor.py
-
+# make service file executable
 chmod +x /home/$user_name/.config/systemd/user/$service_file_name
-
-
-# refresh systemctl and run the service
-systemctl --user daemon-reload
-systemctl --user start $service_file_name
 
 
 # add service to start on startup apps from user ..> to avoid crashing before loading user gui
@@ -80,13 +110,22 @@ if [ ! -d "/home/$user_name/.config/autostart" ];then
         mkdir /home/$user_name/.config/autostart
         echo "Created ~/.config/autostart"
     fi
-cp ExpressVpn_Monitor.desktop /home/$user_name/.config/autostart/
+
+chmod +x ExpressVpn_Monitor.desktop
+cp -a ExpressVpn_Monitor.desktop /home/$user_name/.config/autostart/
 sed -i "4d" ExpressVpn_Monitor.desktop
-cp ExpressVpn_Monitor.desktop /home/$user_name/.local/share/applications/
+cp -a ExpressVpn_Monitor.desktop /home/$user_name/.local/share/applications/
 sed -i "3 a X-GNOME-Autostart-enabled=true" ExpressVpn_Monitor.desktop
 
+sudo chmod +x $app_path_name/*
+sudo chown $user_name $app_path_name/*
+
+# refresh systemctl and run the service
+systemctl --user daemon-reload
+systemctl --user start $service_file_name
+
 # check the service state
-if [[ `systemctl --user is-active $service_file_name` == "active" ]]; then 
-    echo "system `systemctl --user is-active $service_file_name` - Installation Completed .. Done .. BYE :)"; 
+if [[ `systemctl --user is-active $service_file_name` == "active" ]]; then
+    echo "system `systemctl --user is-active $service_file_name` - Installation Completed .. Done .. BYE :)";
 else echo "Installation Failed .... ";
 fi
